+++ e2e/e2e-angular/angular_src/app/app.component.ts
@@ -1,8 +1,11 @@
 import { Component } from '@angular/core';
+import { Homepage } from './src/homepage';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
+  standalone: true,
+  imports: [Homepage],
 })
 export class AppComponent {
   pathname = window.location.pathname;