+++ packages/core/src/__tests__/data/basic-child-component.raw.tsx
@@ -1,5 +1,6 @@
 import { useStore } from '@builder.io/mitosis';
 import MyBasicOnMountUpdateComponent from './basic-onMount-update.raw';
+import MyBasicOutputsComponent from './basic-outputs.raw';
 import MyBasicComponent from './basic.raw';
 
 export default function MyBasicChildComponent() {
@@ -8,11 +9,20 @@ export default function MyBasicChildComponent() {
     dev: 'PatrickJS',
   });
 
+  function log(message: string) {
+    console.log(message);
+  }
+
   return (
     <div>
       <MyBasicComponent id={state.dev} />
       <div>
         <MyBasicOnMountUpdateComponent hi={state.name} bye={state.dev} />
+        <MyBasicOutputsComponent
+          message="Test"
+          onMessageChange={(name) => (state.name = name)}
+          onEvent={() => log('Test')}
+        />
       </div>
     </div>
   );